refactor(CSS): refactor css snapshot resource
